Debug

+++ mistralrs-core/src/sequence.rs
@@ -188,10 +188,10 @@ impl Sequence {
     ) -> Option<StopReason> {
         // TODO(EricLBuehler): Is there a way to avoid this copy?
         let before = Instant::now();
-        let tok = tok.to_scalar::<u8>().unwrap();
+        let tok = tok.to_scalar::<u32>().unwrap();
         println!("Token = {}", before.elapsed().as_millis());
         let before = Instant::now();
-        let eos_tok = eos_tok.to_scalar::<u8>().unwrap();
+        let eos_tok = eos_tok.to_scalar::<u32>().unwrap();
         println!("Token = {}", before.elapsed().as_millis());
         if tok == eos_tok
         {
